+++ backend/tests/backend/main.rs
@@ -4,4 +4,4 @@ mod davs;
 mod davs_litmus;
 mod helpers;
 mod oauth2;
-mod user;
+mod auth;
